old_requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-pyotp==0.6.0
drf-yasg==1.21.5
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.3
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
pyotp==2.8.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
swagger-spec-validator==3.0.3
# twilio==8.2.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
yarl==1.9.2
